Simplify migrations code
